requirements.txt
feedparser
pandas
python-dotenv
streamlit
plotly
requests
beautifulsoup4
lxml
